requirements.txt
django==4.0
psycopg2>=2.8
django-pghistory==1.5.0
djangorestframework==3.13.1